% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominalRR.R
\name{printnRR}
\alias{printnRR}
\title{Print adjusted relative risk under nominal exposure variable.}
\usage{
printnRR(formula, basecov, comparecov, fixcov = NULL, data)
}
\arguments{
\item{formula}{a formula term that is passed into \code{glm()} having a form of \code{response ~ terms} where \code{response} is binary response vector and \code{terms} is a collection of terms connected by \code{'+'}. The first term of predictors will be used as a predictor of interest to calculate relative risks with respect to response variable.}

\item{basecov}{a baseline value of exposure variable. Defaults to the first level.}

\item{comparecov}{a value of exposure variable for comparison. Defaults to the first level.}

\item{fixcov}{a data frame of fixed value for each of adjusted confounders. If there is no confounder other than an exposure variable of interest, \code{fixcov} = \code{NULL}; if \code{fixcov} is missing for covariates, they are all set to \code{0} (for numerical covariates) or first levels (for factor covariates).}

\item{data}{a data frame containing response variable and all the terms used in \code{formula}.}
}
\value{
\item{\code{fit}}{an object of class \code{glm}.}
\item{\code{RR}}{(adjusted) relative risk in response under exposure at baseline (\code{basecov}) and \code{comparecov}.}
\item{\code{delta.var}}{estimated variance of relative risk (\code{RR}) using Delta method.}
\item{\code{fix.cov}}{a data frame of fixed value for each of adjsuted confounders.}
}
\description{
Print adjusted relative risk under nominal exposure variable.
}
\examples{
n <- 500
set.seed(1234)
W <- rbinom(n, 1, 0.3); W[sample(1:n, n/3)] = 2
dat <- as.data.frame(W)
dat$X <- sample( c("low", "medium", "high"), size = n, replace = TRUE)
dat$Y <- ifelse(dat$X == "low", rbinom(n, 1, plogis(W + 0.5)),
               ifelse(dat$X == "medium", rbinom(n, 1, plogis(W + 0.2)),
               rbinom(n, 1, plogis(W - 0.4)) ))
dat$X <- as.factor(dat$X)
result <- printnRR(Y ~ X + W, basecov = "high", comparecov = "low", data = dat)


}
\author{
Youjin Lee
}
